requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.0.3
mido==1.2.9
music21==5.7.2
numpy==1.18.3
protobuf==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
redis==3.5.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==1.12.2
tensorflow-estimator==1.14.0
tensorflow-gpu==1.12.3
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
